Revert "refactor(global): 인증/인가 부분 코틀린 변환  "
