+++ backend/src/main/java/com/backend/domain/auth/service/AuthService.java
@@ -7,8 +7,8 @@ import com.backend.domain.member.entity.Member;
 import com.backend.domain.member.entity.MemberStatus;
 import com.backend.domain.member.repository.MemberRepository;
 import com.backend.global.exception.BusinessException;
-import com.backend.global.security.jwt.JwtTokenProvider;
-import com.backend.global.security.jwt.TokenStatus;
+import com.backend.global.jwt.JwtTokenProvider;
+import com.backend.global.jwt.TokenStatus;
 import com.backend.global.response.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
