--- backend/src/main/java/com/backend/domain/member/entity/MemberStatus.kt
@@ -1,6 +1,0 @@
-package com.backend.domain.member.entity
-
-enum class MemberStatus {
-    ACTIVE,
-    DELETED
-}
