--- backend/src/main/java/com/backend/domain/member/entity/Role.kt
@@ -1,5 +1,0 @@
-package com.backend.domain.member.entity
-
-enum class Role {
-    USER, ADMIN
-}
