--- backend/src/main/java/com/backend/global/config/WebConfig.kt
@@ -1,22 +1,0 @@
-package com.backend.global.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class WebConfig {
-
-    @Bean
-    fun corsConfigurer(): WebMvcConfigurer = object : WebMvcConfigurer {
-        override fun addCorsMappings(registry: CorsRegistry) {
-            registry.addMapping("/**")
-                .allowedOriginPatterns("http://localhost:3000")
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH")
-                .allowCredentials(true)
-                .allowedHeaders("*")
-                .exposedHeaders("Authorization")
-        }
-    }
-}
